Update Jira sync workflow to use pandoc (#5)
* Update Jira sync workflow to use pandoc

* Use a less common EOF marker

A single line in a comment/PR body containing just `EOF` would previously have broken the HereDoc escaping. While uncommon, it wouldn't be that unexpected because we use HereDoc strings a lot in our learning materials for creating Vault policies etc. Now the body would have to contain a single line with just `EOF-PandocInput`, which is hopefully even less common :)